requirements.txt
Django>=4.0
djangorestframework
django-environ
djangorestframework-simplejwt
phonenumbers
django-phonenumber-field
